refactor(quality): split thousand-line CSS, tests, and catalog modules
Keep cascade order with adjacent part imports. Session tests, stall history, and Remote IM schemas drop below the 1000-line gate.
