+++ src/components/lobe-chat/lobe-chat.css
@@ -1,6 +1,7 @@
 /* lobe-chat.css shell */
 @import "./lobe-chat.part1.css";
 @import "./lobe-chat.part2.css";
+@import "./lobe-chat.part2b.css";
 @import "./lobe-chat.part3.css";
 
 /* Plain-stream body retired: live markdown is kept during streaming
