+++ src/lib/reliabilityCenter.test.ts
@@ -1,35 +1,37 @@
 import { describe, expect, it } from "vitest";
 import {
-  applyClearStallHistoryPlan,
   assembleReliabilityCenter,
   buildReliabilityCenter,
-  buildStallHistoryExport,
-  clearStallHistory,
-  buildStallTimelineSnapshot,
   collectLiveStallSignals,
   collectReliabilityBusySessions,
+  mergeErrorEntries,
+  mergeStallSignals,
+  prependReliabilityRing,
+  reliabilityErrorFromDeck,
+  reliabilityStallFromEvent,
+  type ReliabilityErrorEntry,
+  type ReliabilityStallSignal,
+} from "./reliabilityCenter";
+import {
+  applyClearStallHistoryPlan,
+  buildStallHistoryExport,
+  buildStallTimelineSnapshot,
+  clearStallHistory,
   filterStallHistory,
   hasActiveStallHistoryFilters,
   loadStallHistory,
-  mergeErrorEntries,
-  mergeStallSignals,
   parseStallHistory,
   parseStallHistoryEntry,
   planClearStallHistory,
-  prependReliabilityRing,
   recordStallHistory,
   recordStallHistoryFromSignal,
-  reliabilityErrorFromDeck,
-  reliabilityStallFromEvent,
   serializeStallHistoryExport,
-  STALL_HISTORY_MAX,
   serializeStallTimelineSnapshot,
+  STALL_HISTORY_MAX,
   STALL_TIMELINE_FIELD_MAX,
-  type ReliabilityErrorEntry,
-  type ReliabilityStallSignal,
   type StallHistoryEntry,
   type StallHistoryStorage,
-} from "./reliabilityCenter";
+} from "./reliabilityStallHistory";
 import { emptyLiveSnapshot, type SessionLiveMap } from "./sessionLiveStore";
 
 function memStorage(seed?: string): StallHistoryStorage & { store: Map<string, string> } {
