+++ src/styles/chat.css
@@ -1,6 +1,7 @@
 /* chat.css — domain shell */
 @import "./chat.part1.css";
 @import "./chat.part2.css";
+@import "./chat.part2b.css";
 @import "./chat.part3.css";
 @import "./chat.part4.css";
 @import "./chat.part5.css";
