+++ src/styles/composer.css
@@ -1,7 +1,9 @@
 /* composer.css — domain shell */
 @import "./composer.part1.css";
 @import "./composer.part2.css";
+@import "./composer.part2b.css";
 @import "./composer.part3.css";
 @import "./composer.part4.css";
 @import "./composer.part5.css";
 @import "./composer.part6.css";
+@import "./composer.part7.css";
