+++ src/styles/sidebar.css
@@ -1,5 +1,6 @@
 /* sidebar.css — domain shell */
 @import "./sidebar.part1.css";
+@import "./sidebar.part1b.css";
 @import "./sidebar.part2.css";
 @import "./sidebar.part3.css";
 @import "./sidebar.part4.css";
