+++ src/styles/workbench.css
@@ -1,5 +1,6 @@
 /* workbench.css — domain shell */
 @import "./workbench.part1.css";
+@import "./workbench.part1b.css";
 @import "./workbench.part2.css";
 @import "./workbench.part3.css";
 @import "./workbench.part4.css";
